Detach Trashcan into TrashcanMixin, restore -> restore_from_trashcan
