+++ gvm/protocols/gmpv208/__init__.py
@@ -130,6 +130,7 @@ from gvm.protocols.gmpv208.entities.tickets import (
     get_ticket_status_from_string,
 )
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
+from gvm.protocols.gmpv208.entities.trashcan import TrashcanMixin
 from gvm.protocols.gmpv208.entities.user_settings import UserSettingsMixin
 from gvm.protocols.gmpv208.entities.users import (
     UserAuthType,
@@ -168,6 +169,7 @@ class Gmp(
     TasksMixin,
     TicketsMixin,
     TLSCertificateMixin,
+    TrashcanMixin,
     ScanConfigsMixin,
     ScannersMixin,
     SchedulesMixin,
