+++ gvm/protocols/gmpv214/__init__.py
@@ -118,6 +118,7 @@ from gvm.protocols.gmpv208.entities.tickets import (
     get_ticket_status_from_string,
 )
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
+from gvm.protocols.gmpv208.entities.trashcan import TrashcanMixin
 from gvm.protocols.gmpv208.entities.user_settings import UserSettingsMixin
 from gvm.protocols.gmpv208.entities.users import (
     UserAuthType,
@@ -174,6 +175,7 @@ class Gmp(
     TasksMixin,
     TicketsMixin,
     TLSCertificateMixin,
+    TrashcanMixin,
     ScanConfigsMixin,
     ScannersMixin,
     SchedulesMixin,
