+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -26,7 +26,7 @@ from .test_get_system_reports import GmpGetSystemReportsTestCase
 from .test_get_version import GmpGetVersionCommandTestCase
 from .test_help import GmpHelpTestCase
 from .test_modify_auth import GmpModifyAuthTestCase
-from .test_restore import GmpRestoreTestCase
+from .test_restore_from_trashcan import GmpRestoreTestCase
 from .test_sync_cert import GmpSyncCertCommandTestCase
 from .test_sync_scap import GmpSyncScapCommandTestCase
 from .test_with_statement import GmpWithStatementTestCase
